Added checkIfOutputStreamsAreTerminals
It is possible to change the targets of cout and cerr at runtime, for
example by changing their stream buffer or by changing the stdout or
stderr files using system calls. To ensure correctness in all cases,
AEC++ must therefore re-check whether the output stream is connected to
a terminal for every output operation. This would however cause a
significant performance overhead that is almost never needed.

Instead, this commit adds a function "checkIfOutputStreamsAreTerminals",
which can be manually called to re-check whether the standard output
streams are connected to terminals.