Made initialization lazy
Before, initialization was done before the call to main(), in the
constructor of InitializationDummy. This initialization was even
repeated for every translation unit. Now, initialization is done
on-demand, via static initialization inside inline functions. This
removes the overhead at the start of the program.

Also hid a few more functions in the case that the OS is not supported.